added info about template
