+++ package.json
@@ -14,7 +14,8 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
     "recharts": "^2.0.0-beta.1",
-    "redux": "^4.0.5"
+    "redux": "^4.0.5",
+    "typeface-roboto": "0.0.75"
   },
   "scripts": {
     "start": "react-scripts start",
